refactor: Rename source directory to `src`
