+++ .pyre_configuration
@@ -2,7 +2,7 @@
   "strict": true,
   "site_package_search_strategy": "pep561",
   "source_directories": [
-    "."
+    "./src"
   ],
   "search_path": [
     {"site-package": "pyre_extensions"}
